refactor(kafka-explorer): split integration tests into separate files per endpoint
Extract DescribeCluster, ListTopics, DescribeTopic, and DescribeBroker integration tests into dedicated files, keeping only generic connection tests (SASL, SSL, WrongProtocol) in the main file.
