+++ gravitee-apim-rest-api/gravitee-apim-rest-api-kafka-explorer/src/test/java/io/gravitee/rest/api/kafkaexplorer/resource/KafkaExplorerResourceIntegrationTest_DescribeCluster.java
@@ -1,0 +1,84 @@
+/*
+ * Copyright © 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.rest.api.kafkaexplorer.resource;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import io.gravitee.rest.api.kafkaexplorer.rest.model.DescribeClusterRequest;
+import io.gravitee.rest.api.kafkaexplorer.rest.model.DescribeClusterResponse;
+import io.gravitee.rest.api.kafkaexplorer.rest.model.KafkaExplorerError;
+import java.util.Map;
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator;
+import org.junit.jupiter.api.Test;
+
+@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
+class KafkaExplorerResourceIntegrationTest_DescribeCluster extends AbstractKafkaExplorerResourceIntegrationTest {
+
+    @Test
+    void should_return_200_with_cluster_info() {
+        givenClusterWithConfig(Map.of("bootstrapServers", plaintextBootstrapServers(), "security", Map.of("protocol", "PLAINTEXT")));
+
+        var request = new DescribeClusterRequest().clusterId(CLUSTER_ID);
+
+        var response = resource.describeCluster(request);
+
+        assertThat(response.getStatus()).isEqualTo(200);
+        var body = (DescribeClusterResponse) response.getEntity();
+        assertThat(body.getClusterId()).isNotBlank();
+        assertThat(body.getController()).isNotNull();
+        assertThat(body.getNodes()).isNotEmpty();
+        assertThat(body.getTotalTopics()).isGreaterThanOrEqualTo(0);
+        assertThat(body.getTotalPartitions()).isGreaterThanOrEqualTo(0);
+
+        var firstNode = body.getNodes().getFirst();
+        assertThat(firstNode.getHost()).isNotBlank();
+        assertThat(firstNode.getPort()).isGreaterThan(0);
+        assertThat(firstNode.getLeaderPartitions()).isGreaterThanOrEqualTo(0);
+        assertThat(firstNode.getReplicaPartitions()).isGreaterThanOrEqualTo(0);
+    }
+
+    @Test
+    void should_return_400_when_request_is_null() {
+        var response = resource.describeCluster(null);
+
+        assertThat(response.getStatus()).isEqualTo(400);
+        var error = (KafkaExplorerError) response.getEntity();
+        assertThat(error.getTechnicalCode()).isEqualTo("INVALID_PARAMETERS");
+    }
+
+    @Test
+    void should_return_400_when_cluster_id_is_empty() {
+        var request = new DescribeClusterRequest().clusterId("");
+
+        var response = resource.describeCluster(request);
+
+        assertThat(response.getStatus()).isEqualTo(400);
+    }
+
+    @Test
+    void should_return_502_when_broker_unreachable() {
+        givenClusterWithConfig(Map.of("bootstrapServers", "localhost:19092", "security", Map.of("protocol", "PLAINTEXT")));
+
+        var request = new DescribeClusterRequest().clusterId(CLUSTER_ID);
+
+        var response = resource.describeCluster(request);
+
+        assertThat(response.getStatus()).isEqualTo(502);
+        var error = (KafkaExplorerError) response.getEntity();
+        assertThat(error.getTechnicalCode()).isEqualTo("CONNECTION_FAILED");
+    }
+}
